# collection/Norest_Custom_Domain.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Norest Mail - Custom Domain API",
    "description": "Complete API documentation for Norest Mail Custom Domain flow",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
    "/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "User already exists"
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user details"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/domains": {
      "post": {
        "tags": [
          "Custom Domain"
        ],
        "summary": "Add custom domain",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Domain added"
          },
          "400": {
            "description": "Invalid domain"
          },
          "409": {
            "description": "Domain already exists"
          }
        }
      },
      "get": {
        "tags": [
          "Custom Domain"
        ],
        "summary": "List user domains",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of domains"
          }
        }
      }
    },
    "/v1/domains/{id}": {
      "get": {
        "tags": [
          "Custom Domain"
        ],
        "summary": "Get domain details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain details"
          },
          "404": {
            "description": "Domain not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Custom Domain"
        ],
        "summary": "Delete domain",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Domain deleted"
          },
          "404": {
            "description": "Domain not found"
          }
        }
      }
    },
    "/v1/domains/{id}/verification/start": {
      "post": {
        "tags": [
          "Custom Domain"
        ],
        "summary": "Start domain verification",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification started"
          }
        }
      }
    },
    "/v1/domains/{id}/verification": {
      "get": {
        "tags": [
          "Custom Domain"
        ],
        "summary": "Get verification status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification status"
          },
          "404": {
            "description": "Domain not found"
          }
        }
      }
    },
    "/v1/registration/status": {
      "get": {
        "tags": [
          "Registration Flow"
        ],
        "summary": "Get registration status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Registration status"
          }
        }
      }
    },
    "/v1/registration/domains/{domainID}/verify": {
      "post": {
        "tags": [
          "Registration Flow"
        ],
        "summary": "Initiate domain verification",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "domainID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification initiated"
          }
        }
      },
      "get": {
        "tags": [
          "Registration Flow"
        ],
        "summary": "Check domain verification status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "domainID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification status"
          }
        }
      }
    },
...
    "/v1/domains/{domainID}/addresses": {
      "get": {
        "tags": [
          "Email Address"
        ],
        "summary": "List addresses for domain",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "domainID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of addresses"
          }
        }
      }
    },
    "/v1/addresses/{addressID}/claim": {
      "post": {
        "tags": [
          "Email Address"
        ],
        "summary": "Claim reserved address",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "addressID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address claimed successfully"
          },
          "404": {
            "description": "Address not found"
          },
          "409": {
            "description": "Address already claimed"
          }
        }
      }
    },
    "/v1/mail/session": {
      "post": {
        "tags": [
          "Mail Account Provisioning"
        ],
        "summary": "Create mail session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Mail session created"
          },
          "503": {
            "description": "Mailbox not ready"
          }
        }
      }
    },
    "/v1/mail/account": {
      "get": {
        "tags": [
          "Mail Account Provisioning"
        ],
        "summary": "Get mail account details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Account details"
          }
        }
      }
    },
    "/v1/mail/provisioning-status": {
      "get": {
        "tags": [
          "Mail Account Provisioning"
        ],
        "summary": "Get provisioning status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provisioning status"
          }
        }
      }
    },
    "/v1/mail/mailboxes": {
      "get": {
        "tags": [
          "Mailbox"
        ],
        "summary": "List mailboxes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of mailboxes"
          }
        }
      }
    },
    "/v1/mail/mailboxes/{id}": {
      "get": {
        "tags": [
          "Mailbox"
        ],
        "summary": "Get mailbox details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mailbox details"
          },
          "404": {
            "description": "Mailbox not found"
          }
        }
      }
    },
    "/v1/mail/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "List messages",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mailbox_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "position",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subject",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "text",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "thread_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "has_attachment",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages"
          }
        }
      }
    },
    "/v1/mail/messages/{id}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Get message details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message details"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
    "/v1/mail/search": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Search messages",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subject",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mailbox",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "has_attachment",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          }
        }
      }
    },
    "/v1/mail/sync": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Incremental sync",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sync results"
          },
          "409": {
            "description": "Resync required"
          }
        }
      }
    },
    "/v1/mail/attachments": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Upload attachment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment uploaded"
          },
          "413": {
            "description": "Attachment too large"
          }
        }
      }
    },
    "/v1/mail/attachments/{blob_id}": {
      "get": {
        "tags": [
          "Attachments"
        ],
        "summary": "Download attachment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "blob_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attachment file",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Attachment not found"
          }
        }
      }
    },
    "/v1/mail/drafts": {
      "post": {
        "tags": [
          "Drafts"
        ],
        "summary": "Create draft",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "from",
                  "to",
                  "subject"
                ],
                "properties": {
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "cc": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "bcc": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "subject": {
                    "type": "string"
                  },
                  "text_body": {
                    "type": "string"
                  },
                  "html_body": {
                    "type": "string"
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created"
          }
        }
      }
    },
    "/v1/mail/drafts/{id}": {
      "get": {
        "tags": [
          "Drafts"
        ],
        "summary": "Get draft",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Draft details"
          },
          "404": {
            "description": "Draft not found"
          }
        }
      },
      "put": {
        "tags": [
          "Drafts"
        ],
        "summary": "Update draft",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "subject": {
                    "type": "string"
                  },
                  "text_body": {
                    "type": "string"
                  },
                  "html_body": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Draft updated"
          },
          "404": {
            "description": "Draft not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Drafts"
        ],
        "summary": "Delete draft",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Draft deleted"
          },
          "404": {
            "description": "Draft not found"
          }
        }
      }
    },
...
    "/v1/mail/threads": {
      "get": {
        "tags": [
          "Threads"
        ],
        "summary": "List threads",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mailbox_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of threads"
          }
        }
      }
    },
    "/v1/mail/threads/{id}": {
      "get": {
        "tags": [
          "Threads"
        ],
        "summary": "Get thread details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread details"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
...
  }
}